requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.1.8
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.124.2
filelock==3.19.1
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.2
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
open_clip_torch==3.2.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
safetensors==0.7.0
shellingham==1.5.4
starlette==0.49.3
sympy==1.14.0
timm==1.0.22
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wcwidth==0.2.14
